+++ bindings/CXX11/adios2/cxx11/fstream/ADIOS2fstream.cpp
@@ -11,30 +11,8 @@
 #include "ADIOS2fstream.h"
 #include "ADIOS2fstream.tcc"
 
-#ifdef ADIOS2_HAVE_MPI
-#include "adios2/helper/adiosCommMPI.h"
-#endif
-
 namespace adios2
 {
-
-#ifdef ADIOS2_HAVE_MPI
-fstream::fstream(const std::string &name, const openmode mode, MPI_Comm comm,
-                 const std::string engineType)
-: m_Stream(std::make_shared<core::Stream>(
-      name, ToMode(mode), helper::CommFromMPI(comm), engineType, "C++"))
-{
-}
-
-fstream::fstream(const std::string &name, const openmode mode, MPI_Comm comm,
-                 const std::string &configFile,
-                 const std::string ioInConfigFile)
-: m_Stream(std::make_shared<core::Stream>(name, ToMode(mode),
-                                          helper::CommFromMPI(comm), configFile,
-                                          ioInConfigFile, "C++"))
-{
-}
-#endif
 
 fstream::fstream(const std::string &name, const openmode mode,
                  const std::string engineType)
@@ -50,26 +28,6 @@ fstream::fstream(const std::string &name, const openmode mode,
                                           ioInConfigFile, "C++"))
 {
 }
-
-#ifdef ADIOS2_HAVE_MPI
-void fstream::open(const std::string &name, const openmode mode, MPI_Comm comm,
-                   const std::string engineType)
-{
-    CheckOpen(name);
-    m_Stream = std::make_shared<core::Stream>(
-        name, ToMode(mode), helper::CommFromMPI(comm), engineType, "C++");
-}
-
-void fstream::open(const std::string &name, const openmode mode, MPI_Comm comm,
-                   const std::string configFile,
-                   const std::string ioInConfigFile)
-{
-    CheckOpen(name);
-    m_Stream = std::make_shared<core::Stream>(
-        name, ToMode(mode), helper::CommFromMPI(comm), configFile,
-        ioInConfigFile, "C++");
-}
-#endif
 
 void fstream::open(const std::string &name, const openmode mode,
                    const std::string engineType)
